Fix issue with double navigation
Navigation implemented in previous commit actually navigates twice. Not perfect but working for now.
Navigating twice was causing the scene to .open() twice and since the container had no "hidden" class it was
splicing starting at -1 (the last class in the array). This had the odd behavior of clipping off the last class of the container every time you navigated to it.

Still not perfect but improved by only removing "hidden" class if it actually exists.